+++ CHANGELOG.md
@@ -8,8 +8,10 @@
 * Move MCP server to `neocarta` library
 * Change MCP server name to `neocarta-mcp`
 * Update MCP server imports in `eval/` module
+* Deduplicate embedding code in MCP server. MCP server now uses `neocarta` embeddings class.
 
 ### Added
+* Add integration tests for MCP server compatibility with neocarta graph
 
 ## v0.1.0